Removed fpga_utils and made it a part of fpga
